requirements.txt
python-dotenv
tabulate===0.10.0   # For displaying tabular data in a nice format in the console